pyproject.toml
[tool.poetry]
name = "gpug-graphql"
version = "0.1.0"
description = ""
authors = ["Bradley Stuart Kirton <bradleykirton@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.1"
graphene-django = "^2.2"
django-dramatiq = "^0.5.1"
requests = "^2.20"
djangorestframework = "^3.9"
ipython = "^7.1"
djangovue = "^18.11"
dramatiq = {version = "^1.4",extras = ["watch", "redis"]}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
